fix : 매너온도 표시문제 수정

+++ src/main/java/com/jungle/studybbitback/domain/member/dto/FindMemberResponseDto.java
@@ -4,8 +4,8 @@ import com.jungle.studybbitback.domain.dailystudy.entity.DailyStudy;
 import com.jungle.studybbitback.domain.member.entity.Member;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.Duration;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,7 +14,7 @@ public class FindMemberResponseDto {
 	private String email;
 	private String nickname;
 	private String profileImageUrl;
-	private Double flowTemperature;
+	private BigDecimal flowTemperature;
 	private Duration studyTime;
 	private LocalDateTime createdAt;
 
@@ -23,6 +23,7 @@ public class FindMemberResponseDto {
 		this.email = member.getEmail();
 		this.nickname = member.getNickname();
 		this.profileImageUrl = member.getProfileImageUrl();
+		// 소수점 1자리로 설정, 반올림
 		this.flowTemperature = member.getFlowTemperature();
 		if(dailyStudy == null) {
 			this.studyTime = Duration.ofHours(0);
